Support multi-value column type modifiers for `_mysql_data_types_cache` (#126)
Fixes: #122 

This fix will skip over tokens until it finds the closing bracket of the
type modifier `)` instead of consuming exactly 3. This should account
for data types that support more than one value as a modifier, such as:
enum, float, decimal, double.

Additionally, I've added support for enums. There are plugins that use
enum fields and it makes sense to convert those to text in SQLite.

## Testing instructions
Confirm the tests pass