Fix length of index names for a create table statement (#133)
Fixes: #124 

This PR introduces some changes to how we determine the name of an
index. SQLite requires that each index is uniquely named across the
database (not just the table). This causes issues when multiple tables
have the same key names when importing from MySQL. To avoid these
issues, the index name is prefixed with the table name when it is
created. This however results in the indexes having names that are too
long for MySQL so when we run `show create table` the create statements
have these prefixed keys that are too long.

To solve this, we keep adding the table name as a prefix but we remove
it when we generate the create statement. Furthermore, we ensure that
the table name itself has no double underscores in it so we can be sure
that we can split on `__` safely. The result is that each create table
statement will have the original index name and prevents it from being
too long.