Allow double underscores in index names (#136)
See:
https://github.com/WordPress/sqlite-database-integration/pull/133#discussion_r1690678568

When creating new tables in SQLite, original index names are prefixed
with the table name. This is because unlike MySQL, SQLite does not allow
identical index names across the database. Adding a table prefix ensures
the index names are unique.

When running a `show create table` query, we restore the original index
name by removing the table prefix. We split on double underscores but we
did not take into account that the original index name can also contain
`__`. This PR fixes an issue where we explode without a limit. We only
want to explode on the first occurrence of `__`. Unit tests have been
added to ensure this behavior.