Fix the behavior of `HAVING` clause (#154)
The current implementation has a special handling for `HAVING` clauses
without `GROUP BY` (this is [valid in
MySQL](https://stackoverflow.com/questions/6924896/having-without-group-by)
in some SQL modes). This implementation takes advantage of the
equivalence of ungrouped HAVING with WHERE and rewrites such a `HAVING
<condition>` clause to `AND <condition>`.

However, the current implementation has some issues:
1. The presence of `GROUP BY` in a query is incorrectly detected. The
detection always fails, and thus every HAVING query is considered to be
ungrouped. I added a fix in the first commit.
2. Rewriting `HAVING <condition>` to `AND <condition>` fails with an
error `near "AND": syntax error` when no WHERE is included in the query.
This is fixed in the second commit.
3. When an aggregate function is used in HAVING (e.g, `COUNT(*) > 1`),
the query fails with: `misuse of aggregate function COUNT()`. This is
because aggregate functions can't be used in the WHERE clause. This is
fixed in the second commit.

To fix all of these issues, I implemented a fix for the `GROUP BY`
detection (case 1), and changed the rewriting of `HAVING <condition>`
without `GROUP BY` to `GROUP BY 1 HAVING <condition>` (which fixes both
case 2 and 3).