Apply ON UPDATE CURRENT TIMESTAMP in CHANGE queries & fix trigger deletion
Previously, the triger was deleted for all column operations. That was not really correct,
as a side effect of some non-altering operations (e.g, SET/DROP DEFAULT), would then
be removing the trigger.

Additionally, since CHANGE statement handling use continue/break, the trigger creation
was skipped. This commit fixes that and adds a test case for that scenario.