Determine token names lazily, reduce LOC to < 3000, improve lexer performance by ~4%
Token names are used only for debugging purposes.
We can use reflection to determine the names lazily, rather than keeping a full token name map.

This also speeds-up the test suite a bit, which can be due to less lookups and initializations, or because the lexer is initialized for each query.